# web/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jeevanti Hospitals Management System API",
    "description": "Comprehensive API documentation for the Jeevanti Hospitals Management System (HMS) built with Next.js, Prisma, MongoDB, and NextAuth.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/consultations": {
      "get": {
        "summary": "List consultations",
        "parameters": [
          {
            "name": "patientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consultations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consultation"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create consultation",
        "description": "Records clinical consultation (OPD) and updates appointment status to completed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consultation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/prescriptions": {
      "get": {
        "summary": "List prescriptions",
        "description": "Retrieves all prescriptions.",
        "responses": {
          "200": {
            "description": "Prescriptions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prescription"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create prescription",
        "description": "Creates a patient prescription with medicines list.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prescription created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/admissions": {
      "get": {
        "summary": "List admissions",
        "description": "Retrieves a list of inpatient admissions. Optional filter by status.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "admitted",
                "discharged"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Admission"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Admit patient (IPD)",
        "description": "Creates an inpatient admission and marks bed status as occupied.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admission"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/daily-rounds": {
      "get": {
        "summary": "List daily rounds",
        "description": "Retrieves inpatient daily rounds notes.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyRound"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Record daily round",
        "description": "Creates a new daily round record with vitals and doctor clinical notes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyRoundInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Daily round recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyRound"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
    "/api/invoices": {
      "get": {
        "summary": "List invoices",
        "description": "Retrieves a list of invoices.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create invoice",
        "description": "Creates a custom invoice with a set of line items.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/invoices/auto-opd": {
      "post": {
        "summary": "Auto-generate OPD invoice",
        "description": "Auto-generates an invoice for an OPD consultation based on consultation fees, prescribed medicines, and ordered lab tests.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "consultationId"
                ],
                "properties": {
                  "consultationId": {
                    "type": "string",
                    "description": "OPD Consultation ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OPD Invoice auto-generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
    "/api/invoices/auto-ipd": {
      "post": {
        "summary": "Auto-generate IPD invoice",
        "description": "Auto-generates/updates invoice for inpatient stay including room charges, lab tests, and additional services.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "admissionId"
                ],
                "properties": {
                  "admissionId": {
                    "type": "string",
                    "description": "IPD Admission ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPD Invoice auto-generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/lab-orders": {
      "get": {
        "summary": "List lab orders",
        "parameters": [
          {
            "name": "patientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabOrder"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create lab order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabOrderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabOrder"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/lab-tests": {
      "get": {
        "summary": "List lab tests catalog",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabTest"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Add lab test to catalog",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabTestInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabTest"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/wards": {
      "get": {
        "summary": "List wards with bed statistics",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Ward"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "totalBeds": {
                            "type": "integer"
                          },
                          "availableBeds": {
                            "type": "integer"
                          },
                          "occupiedBeds": {
                            "type": "integer"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create ward",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WardInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ward"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/rooms": {
      "post": {
        "summary": "Create room",
        "description": "Creates a room under a ward and optionally auto-creates associated beds.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
    "/api/beds": {
      "get": {
        "summary": "List beds",
        "parameters": [
          {
            "name": "wardId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "available",
                "occupied",
                "maintenance"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bed"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "summary": "List users",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create staff user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/reports": {
      "get": {
        "summary": "Get analytics reports",
        "description": "Retrieves analytical reports of types: daily, monthly, revenue, doctor-performance, or bed-occupancy.",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "daily",
                "monthly",
                "revenue",
                "doctor-performance",
                "bed-occupancy"
              ],
              "default": "daily"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "Date string (YYYY-MM-DD) for daily report",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "query",
            "description": "Month string (YYYY-MM) for monthly/revenue/doctor reports",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "description": "Doctor ID filter for doctor performance report",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response based on report type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid report type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Consultation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "chiefComplaint": {
            "type": "string"
          },
          "symptoms": {
            "type": "string"
          },
          "diagnosis": {
            "type": "string"
          },
          "temperature": {
            "type": "number"
          },
          "bloodPressure": {
            "type": "string"
          },
          "pulse": {
            "type": "integer"
          },
          "respiratoryRate": {
            "type": "integer"
          },
          "oxygenSaturation": {
            "type": "integer"
          },
          "weight": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "clinicalNotes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ConsultationInput": {
        "type": "object",
        "required": [
          "patientId",
          "doctorId"
        ],
        "properties": {
          "appointmentId": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "chiefComplaint": {
            "type": "string"
          },
          "symptoms": {
            "type": "string"
          },
          "diagnosis": {
            "type": "string"
          },
          "temperature": {
            "type": "number"
          },
          "bloodPressure": {
            "type": "string"
          },
          "pulse": {
            "type": "integer"
          },
          "respiratoryRate": {
            "type": "integer"
          },
          "oxygenSaturation": {
            "type": "integer"
          },
          "weight": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "clinicalNotes": {
            "type": "string"
          }
        }
      },
      "Prescription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "consultationId": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "medicines": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "medicineId": {
                  "type": "string"
                },
                "medicineName": {
                  "type": "string"
                },
                "dose": {
                  "type": "string"
                },
                "frequency": {
                  "type": "string"
                },
                "duration": {
                  "type": "string"
                },
                "instructions": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "PrescriptionInput": {
        "type": "object",
        "required": [
          "patientId",
          "medicines"
        ],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "consultationId": {
            "type": "string"
          },
          "medicines": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "medicineId",
                "dosage",
                "frequency",
                "duration"
              ],
              "properties": {
                "medicineId": {
                  "type": "string"
                },
                "dosage": {
                  "type": "string"
                },
                "frequency": {
                  "type": "string"
                },
                "duration": {
                  "type": "string"
                },
                "instructions": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Admission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "wardId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "bedId": {
            "type": "string"
          },
          "admissionNumber": {
            "type": "string"
          },
          "admissionDate": {
            "type": "string",
            "format": "date-time"
          },
          "dischargeDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "admitted",
              "discharged"
            ]
          },
          "dischargeSummary": {
            "type": "string"
          },
          "finalDiagnosis": {
            "type": "string"
          },
          "followUpDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdmissionInput": {
        "type": "object",
        "required": [
          "patientId",
          "doctorId",
          "wardId",
          "roomId",
          "bedId"
        ],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "wardId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "bedId": {
            "type": "string"
          }
        }
      },
...
      "DailyRound": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "admissionId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperature": {
            "type": "number"
          },
          "bloodPressure": {
            "type": "string"
          },
          "pulse": {
            "type": "integer"
          },
          "respiratoryRate": {
            "type": "integer"
          },
          "oxygenSaturation": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          }
        }
      },
      "DailyRoundInput": {
        "type": "object",
        "required": [
          "admissionId",
          "doctorId",
          "notes"
        ],
        "properties": {
          "admissionId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "temperature": {
            "type": "string"
          },
          "bloodPressure": {
            "type": "string"
          },
          "pulse": {
            "type": "string"
          },
          "respiratoryRate": {
            "type": "string"
          },
          "oxygenSaturation": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "400ValidationError": {
        "description": "Validation failed / Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "404NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "500InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}